vision_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
ultralytics==8.1.41
opencv-python-headless==4.9.0.80
pillow>=10.3.0
aiohttp>=3.9.5
pydantic>=2.4.2
python-multipart>=0.0.9
numpy>=1.26.0
